+++ addons/hw_escpos_network_printer/controllers/hw_proxy.py
@@ -39,15 +39,15 @@ class Proxy(http.Controller):
             statuses[driver] = drivers[driver].get_status()
         return statuses
 
-    @http.route('/hw_proxy/hello', type='http', auth='none', cors='*')
+    @http.route("/hw_proxy/hello", type="http", auth="none", cors="*")
     def hello(self):
         return "ping"
 
-    @http.route('/hw_proxy/handshake', type='json', auth='none', cors='*')
+    @http.route("/hw_proxy/handshake", type="json", auth="none", cors="*")
     def handshake(self):
         return True
 
-    @http.route('/hw_proxy/status', type='http', auth='none', cors='*')
+    @http.route("/hw_proxy/status", type="http", auth="none", cors="*")
     def status_http(self, debug=None, **kwargs):
         resp = """
 <!DOCTYPE HTML>
@@ -80,18 +80,25 @@ class Proxy(http.Controller):
 
             status = statuses[driver]
 
-            if status['status'] == 'connecting':
-                color = 'black'
-            elif status['status'] == 'connected':
-                color = 'green'
+            if status["status"] == "connecting":
+                color = "black"
+            elif status["status"] == "connected":
+                color = "green"
             else:
-                color = 'red'
+                color = "red"
 
-            resp += "<h3 style='color:" + color + ";'>" + driver + ' : ' + status[
-                'status'] + "</h3>\n"
+            resp += (
+                "<h3 style='color:"
+                + color
+                + ";'>"
+                + driver
+                + " : "
+                + status["status"]
+                + "</h3>\n"
+            )
             resp += "<ul>\n"
-            for msg in status['messages']:
-                resp += '<li>' + msg + '</li>\n'
+            for msg in status["messages"]:
+                resp += "<li>" + msg + "</li>\n"
             resp += "</ul>\n"
         resp += """
             <h2>Connected Devices</h2>
@@ -99,15 +106,21 @@ class Proxy(http.Controller):
         """
         if debug is None:
             resp += """(<a href="/hw_proxy/status?debug">debug version</a>)"""
-        devices = subprocess.check_output("lsusb").decode('utf-8').split('\n')
+        devices = subprocess.check_output("lsusb").decode("utf-8").split("\n")
         count = 0
         resp += "<div class='devices'>\n"
         for device in devices:
-            device_name = device[device.find('ID') + 2:]
+            device_name = device[device.find("ID") + 2 :]
             if device_name:  # to avoid last empty line
                 device_id = device_name.split()[0]
                 if not (device_id in BANNED_DEVICES):
-                    resp += "<div class='device' data-device='" + device + "'>" + device_name + "</div>\n"
+                    resp += (
+                        "<div class='device' data-device='"
+                        + device
+                        + "'>"
+                        + device_name
+                        + "</div>\n"
+                    )
                     count += 1
 
         if count == 0:
@@ -124,132 +137,114 @@ class Proxy(http.Controller):
                 %s
                 </pre>
 
-            """ % subprocess.check_output('lsusb -v',
-                                          shell=True).decode('utf-8')
+            """ % subprocess.check_output(
+                "lsusb -v", shell=True
+            ).decode(
+                "utf-8"
+            )
 
         return request.make_response(
-            resp, {
-                'Cache-Control': 'no-cache',
-                'Content-Type': 'text/html; charset=utf-8',
-                'Access-Control-Allow-Origin': '*',
-                'Access-Control-Allow-Methods': 'GET',
-            })
+            resp,
+            {
+                "Cache-Control": "no-cache",
+                "Content-Type": "text/html; charset=utf-8",
+                "Access-Control-Allow-Origin": "*",
+                "Access-Control-Allow-Methods": "GET",
+            },
+        )
 
-    @http.route('/hw_proxy/status_json', type='json', auth='none', cors='*')
+    @http.route("/hw_proxy/status_json", type="json", auth="none", cors="*")
     def status_json(self):
         return self.get_status()
 
-    @http.route('/hw_proxy/scan_item_success',
-                type='json',
-                auth='none',
-                cors='*')
+    @http.route("/hw_proxy/scan_item_success", type="json", auth="none", cors="*")
     def scan_item_success(self, ean):
         """
         A product has been scanned with success
         """
-        print('scan_item_success: %s', ean)
+        print("scan_item_success: %s", ean)
 
-    @http.route('/hw_proxy/scan_item_error_unrecognized',
-                type='json',
-                auth='none',
-                cors='*')
+    @http.route(
+        "/hw_proxy/scan_item_error_unrecognized", type="json", auth="none", cors="*"
+    )
     def scan_item_error_unrecognized(self, ean):
         """
         A product has been scanned without success
         """
-        print('scan_item_error_unrecognized: %s', ean)
+        print("scan_item_error_unrecognized: %s", ean)
 
-    @http.route('/hw_proxy/help_needed', type='json', auth='none', cors='*')
+    @http.route("/hw_proxy/help_needed", type="json", auth="none", cors="*")
     def help_needed(self):
         """
         The user wants an help (ex: light is on)
         """
         print("help_needed")
 
-    @http.route('/hw_proxy/help_canceled', type='json', auth='none', cors='*')
+    @http.route("/hw_proxy/help_canceled", type="json", auth="none", cors="*")
     def help_canceled(self):
         """
         The user stops the help request
         """
         print("help_canceled")
 
-    @http.route('/hw_proxy/payment_request',
-                type='json',
-                auth='none',
-                cors='*')
+    @http.route("/hw_proxy/payment_request", type="json", auth="none", cors="*")
     def payment_request(self, price):
         """
         The PoS will activate the method payment
         """
         print("payment_request: price:%s" % price)
-        return 'ok'
+        return "ok"
 
-    @http.route('/hw_proxy/payment_status', type='json', auth='none', cors='*')
+    @http.route("/hw_proxy/payment_status", type="json", auth="none", cors="*")
     def payment_status(self):
         print("payment_status")
-        return {'status': 'waiting'}
+        return {"status": "waiting"}
 
-    @http.route('/hw_proxy/payment_cancel', type='json', auth='none', cors='*')
+    @http.route("/hw_proxy/payment_cancel", type="json", auth="none", cors="*")
     def payment_cancel(self):
         print("payment_cancel")
 
-    @http.route('/hw_proxy/transaction_start',
-                type='json',
-                auth='none',
-                cors='*')
+    @http.route("/hw_proxy/transaction_start", type="json", auth="none", cors="*")
     def transaction_start(self):
-        print('transaction_start')
+        print("transaction_start")
 
-    @http.route('/hw_proxy/transaction_end',
-                type='json',
-                auth='none',
-                cors='*')
+    @http.route("/hw_proxy/transaction_end", type="json", auth="none", cors="*")
     def transaction_end(self):
-        print('transaction_end')
+        print("transaction_end")
 
-    @http.route('/hw_proxy/cashier_mode_activated',
-                type='json',
-                auth='none',
-                cors='*')
+    @http.route("/hw_proxy/cashier_mode_activated", type="json", auth="none", cors="*")
     def cashier_mode_activated(self):
-        print('cashier_mode_activated')
+        print("cashier_mode_activated")
 
-    @http.route('/hw_proxy/cashier_mode_deactivated',
-                type='json',
-                auth='none',
-                cors='*')
+    @http.route(
+        "/hw_proxy/cashier_mode_deactivated", type="json", auth="none", cors="*"
+    )
     def cashier_mode_deactivated(self):
-        print('cashier_mode_deactivated')
+        print("cashier_mode_deactivated")
 
-    @http.route('/hw_proxy/open_cashbox', type='json', auth='none', cors='*')
+    @http.route("/hw_proxy/open_cashbox", type="json", auth="none", cors="*")
     def open_cashbox(self):
-        print('open_cashbox')
+        print("open_cashbox")
 
-    @http.route('/hw_proxy/print_receipt', type='json', auth='none', cors='*')
+    @http.route("/hw_proxy/print_receipt", type="json", auth="none", cors="*")
     def print_receipt(self, receipt):
-        print('print_receipt %s', receipt)
+        print("print_receipt %s", receipt)
 
-    @http.route('/hw_proxy/is_scanner_connected',
-                type='json',
-                auth='none',
-                cors='*')
+    @http.route("/hw_proxy/is_scanner_connected", type="json", auth="none", cors="*")
     def is_scanner_connected(self, receipt):
-        print('is_scanner_connected?')
+        print("is_scanner_connected?")
         return False
 
-    @http.route('/hw_proxy/scanner', type='json', auth='none', cors='*')
+    @http.route("/hw_proxy/scanner", type="json", auth="none", cors="*")
     def scanner(self, receipt):
-        print('scanner')
+        print("scanner")
         time.sleep(10)
-        return ''
+        return ""
 
-    @http.route('/hw_proxy/log', type='json', auth='none', cors='*')
+    @http.route("/hw_proxy/log", type="json", auth="none", cors="*")
     def log(self, arguments):
-        LOGGER.info(' '.join(str(v) for v in arguments))
+        LOGGER.info(" ".join(str(v) for v in arguments))
 
-    @http.route('/hw_proxy/print_pdf_invoice',
-                type='json',
-                auth='none',
-                cors='*')
+    @http.route("/hw_proxy/print_pdf_invoice", type="json", auth="none", cors="*")
     def print_pdf_invoice(self, pdfinvoice):
-        print('print_pdf_invoice %s' % pdfinvoice)
+        print("print_pdf_invoice %s" % pdfinvoice)